+++ src/Core/Exceptions/ContainerException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Niang\Core\Exceptions;
+
+/** Erreur de résolution du conteneur DI : classe introuvable, dépendance circulaire, paramètre manquant. */
+class ContainerException extends \RuntimeException
+{
+}